Couple of feedback fixes
