+++ hydra-plutus/hydra-plutus.cabal
@@ -56,6 +56,7 @@ library
     Hydra.Contract.HeadTokensError
     Hydra.Contract.MintAction
     Hydra.Contract.Util
+    Hydra.Contract.UtilError
     Hydra.Data.ContestationPeriod
     Hydra.Data.Party
     Hydra.Plutus
